Force integer to setLevel for dimmable lights
This should fix #40

+++ smartapps/stj/mqtt-bridge.src/mqtt-bridge.groovy
@@ -640,7 +640,7 @@ def actionColorTemperature(device, attribute, value) {
 }
 
 def actionLevel(device, attribute, value) {
-    device.setLevel(value)
+    device.setLevel(value as int)
 }
 
 def actionConsumable(device, attribute, value) {
